pyproject.toml
[tool.poetry]
name = "grok-to-stl"
version = "0.1.0"
description = "Generate 3D printable models using Grok AI and OpenSCAD"
authors = ["fr0g-66723067 <fr0g@66723067.io>"]
readme = "README.md"
packages = [{include = "grok_openscad", from = "src"}]
repository = "https://github.com/fr0g-66723067/grok-to-stl"

[tool.poetry.dependencies]
python = "^3.8.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
solidpython = "^1.1.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.3.0"
flake8 = "^7.0.0"
tqdm = "^4.65.0"

[tool.poetry.scripts]
grok-to-stl = "grok_openscad.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
